requirements.txt
adal==1.2.7
aiohttp==3.8.1
aiosignal==1.2.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
azure-cognitiveservices-search-imagesearch==2.0.0
azure-common==1.1.28
azure-core==1.24.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
blis==0.7.8
boto==2.49.0
boto3==1.24.20
botocore==1.27.20
catalogue==2.0.7
certifi==2022.6.15
cffi==1.15.0
channels==3.0.3
charset-normalizer==2.1.0
click==8.1.3
constantly==15.1.0
cryptography==37.0.2
cycler==0.11.0
cymem==2.0.6
daphne==3.0.1
datasets==2.3.2
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
Django==3.2.13
django-environ==0.9.0
django-storages==1.12.3
djangorestframework==3.13.1
entrypoints==0.4
executing==0.8.3
fastai==2.7.4
fastbook==0.0.26
fastcore==1.4.5
fastdownload==0.0.6
fastjsonschema==2.15.3
fastprogress==1.0.2
fastrelease==0.1.16
filelock==3.7.1
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
ghapi==0.1.21
graphviz==0.20
gunicorn==20.0.4
huggingface-hub==0.8.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isodate==0.6.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsonschema==4.6.1
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.3
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
msrest==0.7.1
msrestazure==0.6.4
multidict==6.0.2
multiprocess==0.70.13
murmurhash==1.0.7
nbclient==0.6.4
nbconvert==6.5.0
nbdev==1.2.10
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.0
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
Pygments==2.12.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
regex==2022.6.2
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
s3transfer==0.6.0
scikit-learn==1.1.1
scipy==1.8.1
Send2Trash==1.8.0
sentencepiece==0.1.96
service-identity==21.1.0
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
sqlparse==0.4.2
srsly==2.4.3
stack-data==0.3.0
supervisor==4.2.2
terminado==0.15.0
thinc==8.0.17
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
transformers==4.20.1
Twisted==22.4.0
txaio==22.2.1
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
wasabi==0.9.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.1
xxhash==3.0.0
yarl==1.7.2
zope.interface==5.4.0
